Split pipeline into bake & sample
